refactor: colocate reward-store tests
